Puntos 1,2,3,5 hechos

+++ src/trazadorrayos/Esfera.java
@@ -4,18 +4,21 @@
  * and open the template in the editor.
  */
 package trazadorrayos;
+
 import javax.vecmath.Point4d;
+
 /**
  *
  * @author shathe
  */
-public class Esfera extends Figura{
+public class Esfera extends Figura {
 
-    private double radio=0.0;
-    public Esfera(Point4d posicion, Color color, int refraccion, int reflexion, 
-            int kd, int ks, double radio) {
-        super(posicion, color, refraccion, reflexion, kd, ks);
-        this.radio=radio;
-    }
-    
+	private double radio = 0.0;
+
+	public Esfera(Point4d posicion, Color color, double refraccion,
+			double reflexion, double kd, double ks, double radio) {
+		super(posicion, color, refraccion, reflexion, kd, ks);
+		this.radio = radio;
+	}
+
 }
